Re organize to reasonable names
